refactor: Merge gitea package with GiteaClient code
because that package was only used by the GiteaClient code.
