+++ internal/cli/clients/gitea.go
@@ -1,10 +1,13 @@
 package clients
 
 import (
+	"fmt"
+	"strings"
+
 	"code.gitea.io/sdk/gitea"
 	"github.com/pkg/errors"
+	"github.com/suse/carrier/deployments"
 	"github.com/suse/carrier/internal/cli/config"
-	carriergitea "github.com/suse/carrier/internal/gitea"
 	"github.com/suse/carrier/kubernetes"
 )
 
@@ -17,6 +20,16 @@ type GiteaClient struct {
 	Username string
 	Password string
 }
+
+// Resolver figures out where Gitea lives and how to login to it
+type Resolver struct {
+	cluster *kubernetes.Cluster
+	config  *config.Config
+}
+
+const (
+	GiteaCredentialsSecret = "gitea-creds"
+)
 
 var giteaClientMemo *GiteaClient
 
@@ -35,27 +48,26 @@ func GetGiteaClient() (*GiteaClient, error) {
 		return nil, err
 	}
 
-	resolver := carriergitea.NewResolver(configConfig, cluster)
-
-	domain, err := resolver.GetMainDomain()
+	domain, err := getMainDomain(cluster)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to determine carrier domain")
 	}
 
-	giteaURL, err := resolver.GetGiteaURL()
+	giteaURL, err := getGiteaURL(configConfig, cluster)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to resolve gitea host")
 	}
 
-	username, password, err := resolver.GetGiteaCredentials()
+	username, password, err := getGiteaCredentials(cluster)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to resolve gitea credentials")
 	}
 
-	client, err := carriergitea.NewGiteaClient(resolver)
+	client, err := gitea.NewClient(giteaURL)
 	if err != nil {
-		return nil, err
+		return nil, errors.Wrap(err, "gitea client failed")
 	}
+	client.SetBasicAuth(username, password)
 
 	giteaClient := &GiteaClient{
 		Client:   client,
@@ -68,4 +80,82 @@ func GetGiteaClient() (*GiteaClient, error) {
 	giteaClientMemo = giteaClient
 
 	return giteaClient, nil
+}
+
+// getMainDomain finds the main domain for Carrier
+func getMainDomain(cluster *kubernetes.Cluster) (string, error) {
+	// Get the ingress
+	ingresses, err := cluster.ListIngress(deployments.GiteaDeploymentID, "app.kubernetes.io/name=gitea")
+	if err != nil {
+		return "", errors.Wrap(err, "failed to list ingresses for gitea")
+	}
+
+	if len(ingresses.Items) < 1 {
+		return "", errors.New("gitea ingress not found")
+	}
+
+	if len(ingresses.Items) > 1 {
+		return "", errors.New("more than one gitea ingress found")
+	}
+
+	if len(ingresses.Items[0].Spec.Rules) < 1 {
+		return "", errors.New("gitea ingress has no rules")
+	}
+
+	if len(ingresses.Items[0].Spec.Rules) > 1 {
+		return "", errors.New("gitea ingress has more than on rule")
+	}
+
+	host := ingresses.Items[0].Spec.Rules[0].Host
+
+	return strings.TrimPrefix(host, "gitea."), nil
+}
+
+// getGiteaURL finds the URL for gitea
+func getGiteaURL(config *config.Config, cluster *kubernetes.Cluster) (string, error) {
+	// Get the ingress
+	ingresses, err := cluster.ListIngress(deployments.GiteaDeploymentID, "app.kubernetes.io/name=gitea")
+	if err != nil {
+		return "", errors.Wrap(err, "failed to list ingresses for gitea")
+	}
+
+	if len(ingresses.Items) < 1 {
+		return "", errors.New("gitea ingress not found")
+	}
+
+	if len(ingresses.Items) > 1 {
+		return "", errors.New("more than one gitea ingress found")
+	}
+
+	if len(ingresses.Items[0].Spec.Rules) < 1 {
+		return "", errors.New("gitea ingress has no rules")
+	}
+
+	if len(ingresses.Items[0].Spec.Rules) > 1 {
+		return "", errors.New("gitea ingress has more than on rule")
+	}
+
+	host := ingresses.Items[0].Spec.Rules[0].Host
+
+	return fmt.Sprintf("%s://%s", config.GiteaProtocol, host), nil
+}
+
+// getGiteaCredentials resolves Gitea's credentials
+func getGiteaCredentials(cluster *kubernetes.Cluster) (string, string, error) {
+	s, err := cluster.GetSecret(deployments.WorkloadsDeploymentID, GiteaCredentialsSecret)
+	if err != nil {
+		return "", "", errors.Wrap(err, "failed to read gitea credentials")
+	}
+
+	username, ok := s.Data["username"]
+	if !ok {
+		return "", "", errors.Wrap(err, "username key not found in gitea credentials secret")
+	}
+
+	password, ok := s.Data["password"]
+	if !ok {
+		return "", "", errors.Wrap(err, "password key not found in gitea credentials secret")
+	}
+
+	return string(username), string(password), nil
 }
